pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "requests>=2.32.5",
    "xmltodict>=1.0.2",
]
